requirements.txt
numpy
opencv-python
Pillow
loguru 